Move peer node setup to a separate file
